+++ CHANGELOG.md
@@ -6,12 +6,14 @@
     * Added `Utils::simpleTemplate()` method for very simple variable templating
 1. [](#improved)
     * Improved `Flex Users`: obey blueprints and allow Flex to be used in admin only
+    * Improved `Flex` to support custom site template paths
 1. [](#bugfix)
     * Fixed `Page::untranslatedLanguages()` not being symmetrical to `Page::translatedLanguages()`
     * Fixed `Flex Pages` not calling `onPageProcessed` event when cached
     * Fixed phpstan issues in Framework up to level 6
     * Fixed issue with duplicate configuration settings in Flex Directory
     * Fixed fatal error if there are numeric folders in `Flex Pages`
+    * Fixed error on missing `Flex` templates in if `Flex Objects` plugin isn't installed
     * Grav 1.7: Fixed PHP 7.1 compatibility issues
 
 # v1.7.0-beta.10